git-svn-id: http://svn.osgeo.org/osgeo/livedvd/gisvm/trunk@6039 7da46fd8-bd2a-0410-a152-cf9810cf5b10

+++ bin/install_52nWPS.sh
@@ -66,12 +66,18 @@ else
 fi
 
 tar xzf 52n-wps-rc6.tar.gz
-mkdir -p "$INSTALL_FOLDER" --verbose
+
+if [ ! -e "$INSTALL_FOLDER" ] ; then
+   mkdir -p "$INSTALL_FOLDER" --verbose
+fi
+
 
 mv $TMP/52nWPS "$INSTALL_FOLDER"/
-mv $INSTALL_FOLDER/52nWPS/52n.png /usr/share/icons/
 
 
+if [ ! -e "/usr/share/icons/52n.png" ] ; then
+  mv $INSTALL_FOLDER/52nWPS/52n.png /usr/share/icons/
+fi
 
 mkdir -p -v "$USER_HOME/Desktop"
